[스프링 - 협업 미션] 파즈(강승윤) 미션 제출합니다. (#36)
* setting: 설정파일 추가

* refactor: 테이블명 대문자로

* refactor: mysql 포트번호 추가

* refactor: application.yml 삭제

* refactor: ip 변경 및 dml 변경

* refactor: datasource id/pw changed

* feat: application.yml 추가

* feature: 노선목록 조회 API 스펙 변경에 따른 response 값 변경

* feature: email 중복 여부 확인 추가

* feature: 전체보기 신규 조회 API 추가

* refactor: 노선 목록 조회 api 테스트 변경

* feat: 전체보기 조회 신규 API 테스트 작성

* feat: DistanceDiscount 추가

* feat: AgeDiscount 추가

* feat: 요금 계산 추가

* feat: FareTest 추가

* feature: Station 패키지 에러처리

* feat: Line 관련 에러 처리

* feat: Section 관련 에러 처리

* feat: Member 관련 에러 처리

* feat: Auth관련 로직 변경 및 에러 처리

* refactor: api url 'api' 추가

* refactor: cors 설정 config에서 처리

* refactor: Interceptor url 변경

* feat: StationDao 테스트 추가

* feat: MemberDao 테스트 추가

* feat: LineDao 테스트 추가

* refactor: StationService 내부 불필요한 Dao 삭제

* feat: AuthAcceptanceTest 새롭게 추가

* refactor: AcceptanceTest에 default Member 정보 추가

* feature: MemberAcceptanceTest 새롭게 추가

* feat: StationAcceptanceTest 새롭게 추가

* feat: LineAcceptanceTest 새롭게 추가

* refactor: 이전의 acceptanceTest 삭제

* refactor: lombok 설정 변경

* refactor: 노선에 역 등록 되어있으면 역 삭제 안되게끔 추가

* refactor: transactional 설정

* refactor: table name 수정

* refactor: map api,  sectionId출력에서 stationId출력으로 요구사항 변경

* feat: Line 문서화 작업

* feat: AuthController 문서화 작업

* feat: MemberController 문서화

* feat: StationController 문서화

* feat: PathController 문서화

* refactor: prettyPrint() 추가

* refactor: map 요청 오류 수정

* refactor: 테스트 오류 수정

* refactor: 불필요한 import 삭제

* refactor: db info changed

* refactor: error메세지 추출법 변경 및 dao파라미터 전달법 변경

* refactor: Transactional 필요하다 생각하는 곳 추가

* refactor: ControllerTest setUp으로 묶어줌

* refactor: DIstance, Age Enum변경

* refactor: 중복 검사 서비스 단에서 처리

* refactor: 본인 외의 다른 이름을 가진 경우 분기 추가

* refactor: db info changed

* refactor: db port changed

* refactor: db port changed

* refactor: db port changed

* refactor: db port changed

* refactor: db port changed

+++ src/main/java/wooteco/subway/line/application/LineService.java
@@ -1,32 +1,37 @@
 package wooteco.subway.line.application;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.line.dao.LineDao;
 import wooteco.subway.line.dao.SectionDao;
 import wooteco.subway.line.domain.Line;
 import wooteco.subway.line.domain.Section;
-import wooteco.subway.line.dto.LineRequest;
-import wooteco.subway.line.dto.LineResponse;
-import wooteco.subway.line.dto.SectionRequest;
+import wooteco.subway.line.domain.Sections;
+import wooteco.subway.line.dto.*;
+import wooteco.subway.line.exception.line.DuplicatedLineNameException;
+import wooteco.subway.line.exception.line.NoSuchLineException;
+import wooteco.subway.line.exception.section.NoSuchSectionException;
 import wooteco.subway.station.application.StationService;
 import wooteco.subway.station.domain.Station;
 
+import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
+
+import static java.util.stream.Collectors.toList;
 
 @Service
+@RequiredArgsConstructor
 public class LineService {
-    private LineDao lineDao;
-    private SectionDao sectionDao;
-    private StationService stationService;
+    private final LineDao lineDao;
+    private final SectionDao sectionDao;
+    private final StationService stationService;
 
-    public LineService(LineDao lineDao, SectionDao sectionDao, StationService stationService) {
-        this.lineDao = lineDao;
-        this.sectionDao = sectionDao;
-        this.stationService = stationService;
-    }
-
+    @Transactional
     public LineResponse saveLine(LineRequest request) {
+        if (lineDao.isLineExist(request.getName())) {
+            throw new DuplicatedLineNameException();
+        }
         Line persistLine = lineDao.insert(new Line(request.getName(), request.getColor()));
         persistLine.addSection(addInitSection(persistLine, request));
         return LineResponse.of(persistLine);
@@ -42,11 +47,11 @@ public class LineService {
         return null;
     }
 
-    public List<LineResponse> findLineResponses() {
+    public List<TotalLineResponse> findTotalLineResponses() {
         List<Line> persistLines = findLines();
         return persistLines.stream()
-                .map(line -> LineResponse.of(line))
-                .collect(Collectors.toList());
+                .map(TotalLineResponse::of)
+                .collect(toList());
     }
 
     public List<Line> findLines() {
@@ -54,39 +59,101 @@ public class LineService {
     }
 
     public LineResponse findLineResponseById(Long id) {
+        if (lineDao.isLineNotExist(id)) {
+            throw new NoSuchLineException();
+        }
         Line persistLine = findLineById(id);
         return LineResponse.of(persistLine);
     }
 
     public Line findLineById(Long id) {
+        if (lineDao.isLineNotExist(id)) {
+            throw new NoSuchLineException();
+        }
         return lineDao.findById(id);
     }
 
     public void updateLine(Long id, LineRequest lineUpdateRequest) {
+        if (lineDao.isLineNotExist(id)) {
+            throw new NoSuchLineException();
+        }
+        if (lineDao.isLineNameDuplicatedExceptMySelf(id, lineUpdateRequest.getName())) {
+            throw new DuplicatedLineNameException();
+        }
         lineDao.update(new Line(id, lineUpdateRequest.getName(), lineUpdateRequest.getColor()));
     }
 
+    @Transactional
     public void deleteLineById(Long id) {
+        if (lineDao.isLineNotExist(id)) {
+            throw new NoSuchLineException();
+        }
         lineDao.deleteById(id);
+        sectionDao.deleteByLineId(id);
     }
 
+    @Transactional
     public void addLineStation(Long lineId, SectionRequest request) {
         Line line = findLineById(lineId);
         Station upStation = stationService.findStationById(request.getUpStationId());
         Station downStation = stationService.findStationById(request.getDownStationId());
         line.addSection(upStation, downStation, request.getDistance());
 
+        if (lineDao.isLineNotExist(lineId)) {
+            throw new NoSuchSectionException();
+        }
         sectionDao.deleteByLineId(lineId);
         sectionDao.insertSections(line);
     }
 
+    @Transactional
     public void removeLineStation(Long lineId, Long stationId) {
         Line line = findLineById(lineId);
         Station station = stationService.findStationById(stationId);
         line.removeSection(station);
 
+        if (lineDao.isLineNotExist(lineId)) {
+            throw new NoSuchSectionException();
+        }
         sectionDao.deleteByLineId(lineId);
         sectionDao.insertSections(line);
     }
 
+    public List<LineMapResponse> findLineMapResponses() {
+        List<LineMapResponse> lineMapResponses = new ArrayList<>();
+
+        for (Line line : lineDao.findAll()) {
+            Sections sections = line.getSections();
+            List<SectionResponse> sectionResponses = getSectionResponses(sections, line.getId());
+            LineMapResponse lineMapResponse = new LineMapResponse(line, sectionResponses);
+            lineMapResponses.add(lineMapResponse);
+        }
+        return lineMapResponses;
+    }
+
+    private List<SectionResponse> getSectionResponses(Sections sections, Long lineId) {
+        List<SectionResponse> sectionResponses = new ArrayList<>();
+        List<Section> sectionList = sections.getSections();
+
+        for (Section section : sectionList) {
+            List<Line> transferLines = lineDao.findIncludingStation(section.getUpStation().getId(), lineId);
+            List<TransferLineResponse> transferLineResponses = transferLines.stream()
+                    .map(TransferLineResponse::new)
+                    .collect(toList());
+            SectionResponse sectionResponse = new SectionResponse(section, transferLineResponses);
+            sectionResponses.add(sectionResponse);
+        }
+        Section lastSection = sectionList.get(sectionList.size() - 1);
+        sectionResponses.add(getLastSectionResponses(lastSection, lineId));
+        return sectionResponses;
+    }
+
+    private SectionResponse getLastSectionResponses(Section section, Long lineId) {
+        List<Line> transferLines = lineDao.findIncludingStation(section.getDownStation().getId(), lineId);
+        List<TransferLineResponse> transferLineResponses = transferLines.stream()
+                .map(TransferLineResponse::new)
+                .collect(toList());
+        return new SectionResponse(section.getDownStation().getId(), section.getDownStation().getName(),
+                0, transferLineResponses);
+    }
 }
